+++ Assets/Scripts/Runtime/Signals/CoreGameSignals.cs
@@ -10,7 +10,7 @@ namespace Runtime.Signals
     public class CoreGameSignals : MonoSingleton<CoreGameSignals>
     {
         public UnityAction<GameStates> onChangeGameStates = delegate { };
-        public UnityAction<int> onLevelInitialize = delegate { };
+        public UnityAction<byte> onLevelInitialize = delegate { };
         public UnityAction onClearActiveLevel = delegate { };
         public UnityAction onLevelSuccessful = delegate { };
         public UnityAction onLevelFailed = delegate { };
@@ -18,7 +18,7 @@ namespace Runtime.Signals
         public UnityAction onRestartLevel = delegate { };
         public UnityAction onPlay = delegate { };
         public UnityAction onReset = delegate { };
-        public Func<int> onGetLevelID = delegate { return 0; };
+        public Func<byte> onGetLevelID = delegate { return 0; };
         
         public UnityAction onConveyor = delegate {  };
         public UnityAction<GameObject> onAtmTouched = delegate {  };
